add some docs to explain rationale; rename to ExecutionContextProvider
