+++ codepropertygraph/src/main/scala/io/shiftleft/passes/ParallelCpgPass.scala
@@ -1,7 +1,7 @@
 package io.shiftleft.passes
 import io.shiftleft.SerializedCpg
 import io.shiftleft.codepropertygraph.Cpg
-import io.shiftleft.utils.ExecutionContextSummoning
+import io.shiftleft.utils.ExecutionContextProvider
 import org.slf4j.MDC
 
 import java.util.concurrent.LinkedBlockingQueue
@@ -18,7 +18,7 @@ abstract class ParallelCpgPass[T](cpg: Cpg, outName: String = "", keyPools: Opti
 
   def runOnPart(part: T): Iterator[DiffGraph]
 
-  implicit val executionContext: ExecutionContext = ExecutionContextSummoning.summonExecutionContext
+  implicit val executionContext: ExecutionContext = ExecutionContextProvider.getExecutionContext
 
   override def createAndApply(): Unit = {
     withWriter() { writer =>
@@ -35,7 +35,7 @@ abstract class ParallelCpgPass[T](cpg: Cpg, outName: String = "", keyPools: Opti
   private def withWriter[X](serializedCpg: SerializedCpg = new SerializedCpg(),
                             prefix: String = "",
                             inverse: Boolean = false)(f: Writer => Unit): Unit = {
-    val writer = new Writer(serializedCpg, prefix, inverse, MDC.getCopyOfContextMap)
+    val writer = new Writer(serializedCpg, prefix, inverse, MDC.getCopyOfContextMap())
     val writerThread = new Thread(writer)
     writerThread.setName("Writer")
     writerThread.start()
@@ -192,11 +192,11 @@ abstract class ConcurrentWriterCpgPass[T <: AnyRef](cpg: Cpg, outName: String = 
     nParts = parts.size
     val partIter = parts.iterator
     val completionQueue = mutable.ArrayDeque[Future[DiffGraph]]()
-    val writer = new Writer(serializedCpg, prefix, inverse)
+    val writer = new Writer(serializedCpg, prefix, inverse, MDC.getCopyOfContextMap())
     val writerThread = new Thread(writer)
     writerThread.setName("Writer")
     writerThread.start()
-    implicit val ec: ExecutionContext = ExecutionContextSummoning.summonExecutionContext
+    implicit val ec: ExecutionContext = ExecutionContextProvider.getExecutionContext
     try {
       try {
         // The idea is that we have a ringbuffer completionQueue that contains the workunits that are currently in-flight.
@@ -244,7 +244,7 @@ abstract class ConcurrentWriterCpgPass[T <: AnyRef](cpg: Cpg, outName: String = 
   private class Writer(serializedCpg: SerializedCpg,
                        prefix: String,
                        inverse: Boolean,
-                       mdc: java.util.Map[String, String] = MDC.getCopyOfContextMap)
+                       mdc: java.util.Map[String, String])
       extends Runnable {
 
     val queue = new LinkedBlockingQueue[Option[DiffGraph]](ConcurrentWriterCpgPass.writerQueueCapacity)
