+++ codepropertygraph/src/main/scala/io/shiftleft/passes/ParallelIteratorExecutor.scala
@@ -3,7 +3,7 @@ package io.shiftleft.passes
 import scala.concurrent.duration.Duration
 import scala.concurrent.{Await, ExecutionContext, Future}
 
-class ParallelIteratorExecutor[T](iterator: Iterator[T])(implicit val executionContext: ExecutionContext) {
+class ParallelIteratorExecutor[T](iterator: Iterator[T])(implicit executionContext: ExecutionContext) {
   def map[D](func: T => D): Iterator[D] = {
     val futures = Future.traverse(iterator) { element =>
       Future {
